Update functions and inittial menu
